test: address appointment status review feedback
